+++ workflow/scripts/merge_category_monthly_time_series.py
@@ -2,8 +2,8 @@ from collections import defaultdict
 
 from utils import load_json, save_json
 
-result: dict[str, dict[int, list[tuple[int, int, int | float]]]] = defaultdict(
-    lambda: defaultdict(list)
+result: dict[str, dict[str, dict[int, list[tuple[int, int, int | float]]]]] = (
+    defaultdict(lambda: defaultdict(lambda: defaultdict(list)))
 )
 
 for input_file in snakemake.input:
@@ -11,6 +11,7 @@ for input_file in snakemake.input:
 
     for name, data in content.items():
         for aggregation, values in data.items():
-            result[name][aggregation].extend(values)
+            for category, series in values.items():
+                result[name][aggregation][category].extend(series)
 
 save_json(snakemake.output[0], result)