+++ workflow/scripts/strip_format_from_monthly_time_series.py
@@ -4,7 +4,6 @@ data = load_json(snakemake.input[0])
 
 for name, record in data.items():
     if name != "meta":
-        record.pop("styles")
         record.pop("labels")
 
 save_pretty_json(snakemake.output[0], data)