+++ CHANGELOG.md
@@ -5,6 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.2] - 2025-02-06
+
+### Changed
+- Modular metric definition 
+
+### Added
+- Category series.
+- Metric for "Druh nehody"
+- Metric for "Druh srážky jedoucích vozidel"
+- Metric for "Druh pevné překážky"
+- Metric for "Druh zvěře / zvířete"
+- Metric for "Zavinění nehody"
+- Metric for "Alkohol u viníka nehody přítomen"
+- Metric for "Drogy u viníka nehody"
+- Metric for "Hlavní příčiny nehody"
+
+### Removed
+- Normalized and selected months monthly series. Both can be added on display. 
+- Styles.
+
 ## [0.0.1] - 2024-11-06
 
 ### Added